+++ tests/functions.test.js
@@ -1,39 +1,29 @@
-describe("travis", () => {
-  describe("travis", () => {
-      it("should pass travis", () => {
-          expect({status: true}).toEqual({status: true});
-      });
-  });
-});
-
-// test('Checks if the length of the word is less than 10 letters long.', () => {
-//   const werd = new Word();
-//   // Add function to test that word in dictionary has the correct definition
-// });
-
-// const functions = require('./functions');
 
 // test('Adds 2 + 2 to equal 4', () => {
 //   expect(functions.add(2,2)).toBe(4);
 // });
 
+
 // test('Checks "test" word in database for matching definition', () => {
 //   expect('test').toEqual(/these are my notes/);
 // });
 
-// test('Checks "test" word in database for matching definition', () => {
+
+test('Checks "test" word in database for matching definition', () => {
   
-//   const row = {
-//     id: 1,
-//     word: 'word', 
-//     definition: 'definition', 
-//     favorite: '0',
+  class Word {
+    constructor(word_obj) {
+      this.id = 1;
+      this.word = 'helicopter'; 
+      this.definition = 'a type of aircraft which derives both lift and propulsion from one or more sets of horizontally revolving overhead rotors. It is capable of moving vertically and horizontally, the direction of motion being controlled by the pitch of the rotor blades.'; 
+      this.favorite = 0;
+      this.category = 'dirigibles';
+      this.notes = 'helicopters are dangerous!';
+    }
+  }
 
-//     ///
-//   }
 
-//   const word = new Word(row)
+  const werd = new Word();
 
-//   expect(word.word).toBe('test');
-// });
-
+  expect(werd.word).toBe('helicopter');
+});
